requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
aiofiles==23.2.1
opencv-python==4.10.0.84
